Create WeaponFilter component
