+++ libs/gi/page-weapons/src/WeaponFilter.tsx
@@ -1,12 +1,128 @@
-import { SqBadge } from '@genshin-optimizer/common/ui'
+import { CardThemed, SqBadge } from '@genshin-optimizer/common/ui'
+import { catTotal } from '@genshin-optimizer/common/util'
+import { allRarityKeys, allWeaponTypeKeys } from '@genshin-optimizer/gi/consts'
 import type { ICachedWeapon } from '@genshin-optimizer/gi/db'
 import { useDatabase } from '@genshin-optimizer/gi/db-ui'
+import { getWeaponStat } from '@genshin-optimizer/gi/stats'
+import { WeaponRarityToggle, WeaponToggle } from '@genshin-optimizer/gi/ui'
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever'
 import LockIcon from '@mui/icons-material/Lock'
 import LockOpenIcon from '@mui/icons-material/LockOpen'
-import { Button, Grid, Typography } from '@mui/material'
-import { useMemo } from 'react'
+import {
+  Box,
+  Button,
+  CardContent,
+  Grid,
+  TextField,
+  Typography,
+} from '@mui/material'
+import type { ChangeEvent, Dispatch, SetStateAction } from 'react'
+import { useEffect, useMemo, useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
+
+function WeaponFilter({
+  numShowing,
+  total,
+  weaponIds,
+  searchTerm,
+  setSearchTerm,
+}: {
+  numShowing: number
+  total: number
+  weaponIds: string[]
+  searchTerm: string
+  setSearchTerm: Dispatch<SetStateAction<string>>
+}) {
+  const { t } = useTranslation(['page_weapon', 'ui'])
+  const database = useDatabase()
+  const [state, setState] = useState(database.displayWeapon.get())
+
+  useEffect(() => {
+    database.displayWeapon.follow((r, dbMeta) => setState(dbMeta))
+  }, [database])
+
+  const { weaponType, rarity } = state
+
+  const weaponTotals = useMemo(
+    () =>
+      catTotal(allWeaponTypeKeys, (ct) =>
+        database.weapons.entries.forEach(([id, weapon]) => {
+          const wtk = getWeaponStat(weapon.key).weaponType
+          ct[wtk].total++
+          if (weaponIds.includes(id)) ct[wtk].current++
+        })
+      ),
+    [database, weaponIds]
+  )
+
+  const weaponRarityTotals = useMemo(
+    () =>
+      catTotal(allRarityKeys, (ct) =>
+        database.weapons.entries.forEach(([id, weapon]) => {
+          const wr = getWeaponStat(weapon.key).rarity
+          ct[wr].total++
+          if (weaponIds.includes(id)) ct[wr].current++
+        })
+      ),
+    [database, weaponIds]
+  )
+
+  return (
+    <CardThemed>
+      <CardContent sx={{ display: 'flex', flexDirection: 'column', gap: 1 }}>
+        <Grid container>
+          <Grid item>
+            <Typography variant="h6">
+              <Trans t={t} i18nKey="weaponFilter">
+                Weapon Filter
+              </Trans>
+            </Typography>
+          </Grid>
+          <Grid
+            item
+            flexGrow={1}
+            display="flex"
+            justifyContent="center"
+            alignItems="center"
+          >
+            <Typography>
+              <strong>{numShowing}</strong> / {total}
+            </Typography>
+          </Grid>
+        </Grid>
+        <Box display="flex" flexWrap="wrap" gap={1} alignItems="stretch">
+          <WeaponToggle
+            onChange={(weaponType) =>
+              database.displayWeapon.set({ weaponType })
+            }
+            value={weaponType}
+            totals={weaponTotals}
+            size="small"
+          />
+          <WeaponRarityToggle
+            sx={{ height: '100%' }}
+            onChange={(rarity) => database.displayWeapon.set({ rarity })}
+            value={rarity}
+            totals={weaponRarityTotals}
+            size="small"
+          />
+          <Box flexGrow={1} />
+          <TextField
+            autoFocus
+            size="small"
+            value={searchTerm}
+            onChange={(e: ChangeEvent<HTMLTextAreaElement>) =>
+              setSearchTerm(e.target.value)
+            }
+            label={t('weaponName')}
+            sx={{ height: '100%' }}
+            InputProps={{ sx: { height: '100%' } }}
+          />
+        </Box>
+      </CardContent>
+    </CardThemed>
+  )
+}
 
 function WeaponRedButtons({ weaponIds }: { weaponIds: string[] }) {
   const { t } = useTranslation(['weapon', 'ui'])
@@ -104,4 +220,5 @@ function WeaponRedButtons({ weaponIds }: { weaponIds: string[] }) {
   )
 }
 
+export default WeaponFilter
 export { WeaponRedButtons }
